TINY-7191: Moved PositionUtils
